+++ src/EggDotNet/Format/Alz/Header.cs
@@ -1,14 +1,14 @@
-﻿using EggDotNet.Extensions;
+﻿using EggDotNet.InternalExtensions;
 using System;
 using System.IO;
 
 #if NETSTANDARD2_0
-using BitConverter = EggDotNet.Extensions.BitConverterWrapper;
+using BitConverter = EggDotNet.InternalExtensions.BitConverterWrapper;
 #endif
 
 namespace EggDotNet.Format.Alz
 {
-	internal sealed class Header
+    internal sealed class Header
 	{
 		public const int ALZ_HEADER_MAGIC = 0x015A4C41;
