+++ src/EggDotNet/Format/Egg/EncryptHeader.cs
@@ -1,14 +1,14 @@
-﻿using EggDotNet.Extensions;
+﻿using EggDotNet.InternalExtensions;
 using System;
 using System.IO;
 
 #if NETSTANDARD2_0
-using BitConverter = EggDotNet.Extensions.BitConverterWrapper;
+using BitConverter = EggDotNet.InternalExtensions.BitConverterWrapper;
 #endif
 
 namespace EggDotNet.Format.Egg
 {
-	internal sealed class EncryptHeader
+    internal sealed class EncryptHeader
 	{
 		public const int EGG_ENCRYPT_HEADER_MAGIC = 0x08D1470F;
