+++ src/EggDotNet/Format/Egg/FilenameHeader.cs
@@ -1,17 +1,17 @@
 ﻿using EggDotNet.Exceptions;
-using EggDotNet.Extensions;
+using EggDotNet.InternalExtensions;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Text;
 
 #if NETSTANDARD2_0
-using BitConverter = EggDotNet.Extensions.BitConverterWrapper;
+using BitConverter = EggDotNet.InternalExtensions.BitConverterWrapper;
 #endif
 
 namespace EggDotNet.Format.Egg
 {
-	internal sealed class FilenameHeader //: ExtraField2
+    internal sealed class FilenameHeader //: ExtraField2
 	{
 		[Flags]
 		public enum FilenameFlags
