+++ src/EggDotNet/Format/EggFileFormatFactory.cs
@@ -1,10 +1,12 @@
 ﻿using EggDotNet.Exceptions;
-using EggDotNet.Extensions;
-using EggDotNet.Format;
-using System.Collections.Generic;
 using System;
 using System.IO;
 using static EggDotNet.Callbacks;
+
+#if NETSTANDARD2_0
+using EggDotNet.InternalExtensions;
+using BitConverter = EggDotNet.InternalExtensions.BitConverterWrapper;
+#endif
 
 namespace EggDotNet.Format
 {
@@ -14,32 +16,24 @@ namespace EggDotNet.Format
 		{
 #if NETSTANDARD2_1_OR_GREATER
 			Span<byte> headerBuffer = stackalloc byte[Global.HEADER_SIZE];
+#else
+			var headerBuffer = new byte[Global.HEADER_SIZE];
+#endif
 			if (stream.Read(headerBuffer) != Global.HEADER_SIZE)
 			{
 				throw new InvalidDataException("Failed reading archive header");
 			}
 
 			var header = BitConverter.ToInt32(headerBuffer);
-#else
-			var headerBuffer = new byte[Global.HEADER_SIZE];
-			if (stream.Read(headerBuffer, 0, Global.HEADER_SIZE) != Global.HEADER_SIZE)
-			{
-				throw new InvalidDataException("Failed reading archive header");
-			}
 
-			var header = BitConverter.ToInt32(headerBuffer, 0);
-#endif
-			if (header == Egg.Header.EGG_HEADER_MAGIC)
+			switch (header)
 			{
-				return new Egg.EggFormat(streamCallback, pwCallback);
-			}
-			else if (header == Alz.Header.ALZ_HEADER_MAGIC)
-			{
-				return new Alz.AlzFormat();
-			}
-			else
-			{
-				throw new UnknownEggException();
+				case Egg.Header.EGG_HEADER_MAGIC:
+					return new Egg.EggFormat(streamCallback, pwCallback);
+				case Alz.Header.ALZ_HEADER_MAGIC:
+					return new Alz.AlzFormat();
+				default:
+					throw new UnknownEggException();
 			}
 		}
 	}
