+++ src/EggDotNet/InternalExtensions/StreamExtensions.cs
@@ -1,0 +1,14 @@
+﻿using System.IO;
+
+namespace EggDotNet.InternalExtensions
+{
+#if NETSTANDARD2_0
+    internal static class StreamExtensions
+    {
+        public static int Read(this Stream stream, byte[] buffer)
+        {
+            return stream.Read(buffer, 0, buffer.Length);
+        }
+    }
+#endif
+}
